changes in pthread management
git-svn-id: https://femdk.l5.pk.edu.pl/bstools/BSTools/trunk@113 4ce36bd9-3a78-4973-a166-986cbda2e52d

+++ src/pkvaria/scratchmem.c
@@ -78,7 +78,7 @@ static void *_pkv_GetScratchMemTop ( void )
 
 static void _pkv_SetScratchMemTop ( void *p )
 {
-  int fsm;
+  long int fsm;
 
   fsm = (char*)p - FreeScratch;
   FreeScratch = (char*)p;
